+++ src/components/stakingPoolsList.tsx
@@ -3,6 +3,7 @@ import StakingPoolCard from "./stakingPoolCard"
 import SortBtn from "./sortBtn"
 import { useEffect, useMemo, useState } from "react"
 import { StakingPoolType } from "@/misc/stakingPoolsConfig"
+import FastFadeScroll from "@/components/FastFadeScroll"
 
 const StakingPoolsList: React.FC = () => {
   const {
@@ -71,21 +72,7 @@ const StakingPoolsList: React.FC = () => {
       type: StakingPoolType.NORMAL,
     },
   ]
-  const [isScrolling, setIsScrolling] = useState(false)
-  let scrollTimeout: any
 
-  const handleScroll = () => {
-    setIsScrolling(true)
-    clearTimeout(scrollTimeout)
-
-    scrollTimeout = setTimeout(() => {
-      setIsScrolling(false)
-    }, 1000)
-  }
-
-  useEffect(() => {
-    return () => clearTimeout(scrollTimeout)
-  }, [])
   return (
     <>
       <nav
@@ -126,10 +113,7 @@ const StakingPoolsList: React.FC = () => {
           />
         </div>
 
-        <div
-          onScroll={handleScroll}
-          className={`flex-1 pb-4 mb-16 md:mb-0 overflow-y-scroll ${isScrolling ? "scrollbar-visible" : "scrollbar-hidden"} scrollbar-gradient `}
-        >
+        <FastFadeScroll className="flex-1 pb-4 mb-16 md:mb-0 overflow-y-scroll">
           <div className="grid grid-cols-1 gap-2.5 lg:gap-4 4k:gap-5">
             {sortedLiquidStakingPoolsData.map(({ stakingPool, userData }) => (
               <StakingPoolCard
@@ -146,7 +130,7 @@ const StakingPoolsList: React.FC = () => {
               />
             ))}
           </div>
-        </div>
+        </FastFadeScroll>
       </>
     </>
   )
